+++ CHANGES.rst
@@ -9,6 +9,8 @@
   (`Issue #97 <https://github.com/sqlalchemy-redshift/sqlalchemy-redshift/pull/97>`_)
 - Performance improvement for reflection by caching table constraint info
   (`Issue #101 <https://github.com/sqlalchemy-redshift/sqlalchemy-redshift/issues/101>`_)
+- Support BZIP2 compression in COPY command
+  (`Issue #110 <https://github.com/sqlalchemy-redshift/sqlalchemy-redshift/issues/110>`_)
 
 
 0.5.0 (2016-04-21)